Controller Fixes

+++ service-booking/src/main/java/com/example/service/booking/repository/BookingRepository.java
@@ -1,12 +1,11 @@
 package com.example.service.booking.repository;
 
-
 import com.example.service.booking.entities.Booking;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookingRepository extends JpaRepository<Booking, Long> {
+public interface BookingRepository extends JpaRepository<Booking, Integer> {
 
 }
